+++ spec/specification/set_spec.rb
@@ -26,23 +26,6 @@ module Pod
         @set.highest_version_spec_path.should == @source.repo + 'CocoaLumberjack/1.6.2/CocoaLumberjack.podspec'
       end
 
-      it 'checks if the dependency of the specification is compatible with existing requirements' do
-        @set.required_by(Dependency.new('CocoaLumberjack', '1.2'), 'Spec')
-        @set.required_by(Dependency.new('CocoaLumberjack', '< 1.2.1'), 'Spec')
-        @set.required_by(Dependency.new('CocoaLumberjack', '> 1.1'), 'Spec')
-        @set.required_by(Dependency.new('CocoaLumberjack', '~> 1.2.0'), 'Spec')
-        @set.required_by(Dependency.new('CocoaLumberjack'), 'Spec')
-        lambda do
-          @set.required_by(Dependency.new('CocoaLumberjack', '< 1.0'), 'Spec')
-        end.should.raise Informative
-      end
-
-      it "raises if the required version doesn't exist" do
-        should.raise Informative do
-          @set.required_by(Dependency.new('CocoaLumberjack', '< 1.0'), 'Spec')
-        end
-      end
-
       it 'can test if it is equal to another set' do
         @set.should == Spec::Set.new('CocoaLumberjack', @source)
         @set.should.not == Spec::Set.new('RestKit', @source)
@@ -65,41 +48,11 @@ module Pod
 
       #--------------------------------------#
 
-      before do
-        @set.required_by(Dependency.new('CocoaLumberjack', '< 1.2.1'), 'Spec')
-      end
-
-      it 'returns the version required for the dependency' do
-        @set.required_version.should == Version.new('1.2')
-      end
-
-      it 'returns the acceptable versions according to the requirements stored' do
-        @set.acceptable_versions.map(&:to_s).should == ['1.2', '1.1', '1.0']
-      end
-
-      it 'returns the specification for the required version' do
-        @set.specification.name.should == 'CocoaLumberjack'
-        @set.specification.version.should == Version.new('1.2')
-      end
-
       it 'ignores dot files when getting the version directories' do
         `touch #{fixture('spec-repos/master/CocoaLumberjack/.DS_Store')}`
         should.not.raise do
           @set.versions
         end
-      end
-
-      it 'raises if a version is incompatible with the activated version' do
-        dep = Dependency.new('CocoaLumberjack', '1.2.1')
-        should.raise Informative do
-          @set.required_by(dep, 'Spec')
-        end
-      end
-
-      it 'accepts a requirement if it allows supported versions' do
-        dep = Dependency.new('CocoaLumberjack', '< 1.1')
-        @set.required_by(dep, 'Spec')
-        @set.acceptable_versions.map(&:to_s).should == ['1.0']
       end
     end
 
@@ -128,33 +81,6 @@ module Pod
         hash['test_repo'].should == %w(999.999.999 1.13 1.4)
         hash.keys.sort.should == %w(master test_repo)
       end
-
-      it 'returns the specification from the `master` source for the required version' do
-        dep = Dependency.new('JSONKit', '1.5pre')
-        @set.required_by(dep, 'Spec')
-        spec = @set.specification
-        spec.name.should == 'JSONKit'
-        spec.version.to_s.should == '1.5pre'
-        spec.defined_in_file.should == fixture('spec-repos/master/JSONKit/1.5pre/JSONKit.podspec')
-      end
-
-      it 'returns the specification from `test_repo` source for the required version' do
-        dep = Dependency.new('JSONKit', '999.999.999')
-        @set.required_by(dep, 'Spec')
-        spec = @set.specification
-        spec.name.should == 'JSONKit'
-        spec.version.to_s.should == '999.999.999'
-        spec.defined_in_file.should == fixture('spec-repos/test_repo/Specs/JSONKit/999.999.999/JSONKit.podspec')
-      end
-
-      it 'prefers sources by the order in which they were provided' do
-        dep = Dependency.new('JSONKit', '1.4')
-        @set.required_by(dep, 'Spec')
-        spec = @set.specification
-        spec.name.should == 'JSONKit'
-        spec.version.to_s.should == '1.4'
-        spec.defined_in_file.should ==  fixture('spec-repos/test_repo/Specs/JSONKit/1.4/JSONKit.podspec')
-      end
     end
   end
 
@@ -181,23 +107,6 @@ module Pod
 
     it 'returns the version of the specification' do
       @set.versions.map(&:to_s).should == ['1.0']
-    end
-
-    it "doesn't nil the initialization specification on #required_by" do
-      @set.required_by(Dependency.new('BananaLib', '1.0'), 'Spec')
-      @set.specification.should == @spec
-    end
-
-    it 'raises if asked for the specification path' do
-      should.raise StandardError do
-        @set.specification_path
-      end
-    end
-
-    it "raises if the required version doesn't match the specification" do
-      should.raise Informative do
-        @set.required_by(Dependency.new('BananaLib', '< 1.0'), 'Spec')
-      end
     end
   end
 end